refactor(app): relocate Runtime + main to runtime/; thin app.py (#80)
Capstone of decomposing src/vanchor/app.py (epic #68). Moves the Runtime
class into runtime/runtime.py and main() into runtime/cli.py, extracts the
shared module-level constants into runtime/constants.py, and reduces app.py
to a thin back-compat facade that re-exports the entire historical public
surface (Runtime, main, apply_demo_mode, demo_route_waypoints, _TeeMotor,
the failsafe constants, and the previously module-level symbols).

- Repoint collaborator back-imports off ..app: safety_runtime._underway now
  imports the failsafe constants from .constants; ui/server.py's TYPE_CHECKING
  Runtime import points at ..runtime.runtime. No collaborator imports ..app,
  so no cycle (app -> runtime.runtime -> collaborators).
- Drop unused module-level `import math` / `import tempfile` (a local
  `import tempfile as _tf` remains where needed).
- Remove redundant apply_device_overrides import in sessions.restore_backup's
  boat-profiles try-block (the unconditional import later covers it).
- Docs: rewrite backend.md's runtime section to describe runtime/ package.

Entry point unchanged: `vanchor = "vanchor.app:main"` resolves via the app.py
re-export. Full suite green (2137 passed), ruff clean.
